move uploading method to another service
