+++ driver-service/src/main/java/org/modsen/service/FileService.java
@@ -1,0 +1,12 @@
+package org.modsen.service;
+
+import org.modsen.exception.RequestTimeoutException;
+
+import java.io.InputStream;
+
+public interface FileService {
+
+    String uploadFile(String bucketName, String objectName, InputStream inputStream, String contentType)
+            throws RequestTimeoutException;
+
+}
