+++ driver-service/src/main/java/org/modsen/service/impl/FileServiceImpl.java
@@ -1,0 +1,45 @@
+package org.modsen.service.impl;
+
+import io.minio.BucketExistsArgs;
+import io.minio.MakeBucketArgs;
+import io.minio.MinioClient;
+import io.minio.PutObjectArgs;
+import java.io.InputStream;
+import lombok.RequiredArgsConstructor;
+import org.modsen.exception.RequestTimeoutException;
+import org.modsen.service.FileService;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class FileServiceImpl implements FileService {
+
+    private final MinioClient minioClient;
+
+    @Override
+    public String uploadFile(String bucketName, String objectName, InputStream inputStream, String contentType)
+        throws RequestTimeoutException {
+        try {
+            boolean found = minioClient.bucketExists(BucketExistsArgs.builder()
+                .bucket(bucketName)
+                .build());
+            if (!found) {
+                minioClient.makeBucket(MakeBucketArgs.builder()
+                    .bucket(bucketName)
+                    .build());
+            }
+            minioClient.putObject(
+                PutObjectArgs.builder()
+                    .bucket(bucketName)
+                    .object(objectName)
+                    .stream(inputStream, inputStream.available(), -1)
+                    .contentType(contentType)
+                    .build()
+            );
+            return objectName;
+        } catch (Exception e) {
+            throw new RequestTimeoutException(e.getMessage());
+        }
+    }
+
+}
