+++ passenger-service/src/main/java/org/modsen/service/impl/FileService.java
@@ -1,0 +1,11 @@
+package org.modsen.service.impl;
+
+import java.io.InputStream;
+import org.modsen.exception.RequestTimeoutException;
+
+public interface FileService {
+
+    String uploadFile(String bucketName, String objectName, InputStream inputStream, String contentType)
+        throws RequestTimeoutException;
+
+}
