+++ passenger-service/src/main/java/org/modsen/service/impl/StorageServiceImpl.java
@@ -1,11 +1,8 @@
 package org.modsen.service.impl;
 
-import io.minio.BucketExistsArgs;
-import io.minio.MakeBucketArgs;
-import io.minio.MinioClient;
-import io.minio.PutObjectArgs;
+import static org.modsen.util.constant.PassengerServiceConstant.PHOTO_PASSENGER_NAME;
+
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.modsen.config.MinioConfigProperties;
@@ -19,42 +16,15 @@ import org.springframework.web.multipart.MultipartFile;
 @RequiredArgsConstructor
 public class StorageServiceImpl implements StorageService {
 
-    public final static String PHOTO_PASSENGER_NAME = "passenger_photo_%s";
-    private final MinioClient minioClient;
     private final MinioConfigProperties properties;
+
     private final PassengerService passengerService;
 
-    @Override
-    public String uploadFile(String bucketName, String objectName, InputStream inputStream, String contentType)
-        throws RequestTimeoutException {
-        try {
-            boolean found = minioClient.bucketExists(
-                BucketExistsArgs.builder()
-                    .bucket(bucketName)
-                    .build());
-            if (!found) {
-                minioClient.makeBucket(MakeBucketArgs
-                    .builder()
-                    .bucket(bucketName)
-                    .build());
-            }
-            minioClient.putObject(
-                PutObjectArgs.builder()
-                    .bucket(bucketName)
-                    .object(objectName)
-                    .stream(inputStream, inputStream.available(), -1)
-                    .contentType(contentType)
-                    .build()
-            );
-            return objectName;
-        } catch (Exception e) {
-            throw new RequestTimeoutException(e.getMessage());
-        }
-    }
+    private final FileService fileService;
 
     @Override
     public UUID saveFileReference(MultipartFile photoFile, UUID id) throws IOException, RequestTimeoutException {
-        String fileRef = uploadFile(properties.bucket().getPhotoBucketName(),
+        String fileRef = fileService.uploadFile(properties.bucket().getPhotoBucketName(),
             PHOTO_PASSENGER_NAME.formatted(id), photoFile.getInputStream(),
             photoFile.getContentType());
         return passengerService.addPhoto(id, fileRef);
